Create GO_REF:0000112
Request for GO_REF for  in silico-based, curator reviewed GO annotations in the context of the Community Assessment of Community Annotation with Ontologies (CACAO). The intent of this GO reference is to provide detailed criteria for the use of computational methods to assign function/process/component attributes to gene products in published genomes, allowing curators participating in CACAO to perform annotations based on the results of such methods using the GO_REF as a source. The specific criteria and rationales used for each computational method are described and versioned in the CACAO GO_REF category (http://gowiki.tamu.edu/wiki/index.php/Category:CACAO_GO_REF).