Update GO_REF schema to reflect agreed upon format
These changes add a `description` slot to hold the main text of the GO_REF. The `comments` slot is now reserved for the rare instances where a GO_REF contains additional comments presented outside the main description. Other slots remain largely the same except for adding some additional validation to some of them. Finally, the `GoRefCollection` collection is removed as we will use a top-level array in the `gorefs.yaml` data file.